+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/Events/MeteringPointCreated.cs
@@ -16,7 +16,7 @@ using System;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using NodaTime;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption
+namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Events
 {
     public class MeteringPointCreated : DomainEventBase
     {
