+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringPoints/Consumption/CreationTests.cs
@@ -22,7 +22,7 @@ using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints;
 using Xunit;
 using Xunit.Categories;
-using MeteringPointCreated = Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption.MeteringPointCreated;
+using MeteringPointCreated = Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Events.MeteringPointCreated;
 
 namespace Energinet.DataHub.MeteringPoints.Tests.Domain.MeteringPoints.Consumption
 {
